+++ src/main/java/gtexpert/common/metatileentities/EIOMetaTileEntities.java
@@ -7,7 +7,7 @@ import static gtexpert.common.metatileentities.MetaTileEntitiesManager.registerG
 
 import gregtech.api.util.GTUtility;
 
-import gtexpert.api.recipes.GTERecipeMaps;
+import gtexpert.core.GTERecipeMaps;
 import gtexpert.api.util.GTEUtility;
 import gtexpert.client.GTETextures;
 import gtexpert.common.metatileentities.single.electric.MetaTileEntityElectricSpawner;
