+++ src/main/java/gtexpert/common/metatileentities/multi/MetaTileEntitySawmill.java
@@ -26,7 +26,7 @@ import gregtech.client.renderer.ICubeRenderer;
 import gregtech.core.sound.GTSoundEvents;
 
 import gtexpert.api.gui.GTEGuiTextures;
-import gtexpert.api.recipes.GTERecipeMaps;
+import gtexpert.core.GTERecipeMaps;
 import gtexpert.client.GTETextures;
 import gtexpert.common.blocks.GTEBlockMetalCasing;
 import gtexpert.common.blocks.GTEMetaBlocks;
