+++ src/main/java/gtexpert/common/metatileentities/single/steam/SteamMixer.java
@@ -24,7 +24,7 @@ import gregtech.api.metatileentity.interfaces.IGregTechTileEntity;
 import gregtech.client.renderer.texture.Textures;
 
 import gtexpert.api.gui.GTEGuiTextures;
-import gtexpert.api.recipes.GTERecipeMaps;
+import gtexpert.core.GTERecipeMaps;
 
 public class SteamMixer extends SteamMetaTileEntity {
