+++ src/main/java/gtexpert/integration/deda/recipemaps/RecipeMapDraconicUpgrade.java
@@ -1,4 +1,4 @@
-package gtexpert.api.recipes.draconic;
+package gtexpert.integration.deda.recipemaps;
 
 import org.jetbrains.annotations.NotNull;
