+++ src/main/java/gtexpert/integration/deda/recipemaps/upgrade/UpgradeRecipeBuilder.java
@@ -1,4 +1,4 @@
-package gtexpert.api.recipes.draconic.upgrade;
+package gtexpert.integration.deda.recipemaps.upgrade;
 
 import net.minecraft.item.ItemStack;
