+++ src/main/java/gtexpert/integration/deda/recipemaps/upgrade/UpgradeRecipeProperty.java
@@ -1,4 +1,4 @@
-package gtexpert.api.recipes.draconic.upgrade;
+package gtexpert.integration.deda.recipemaps.upgrade;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.I18n;
