+++ src/main/java/gtexpert/integration/deda/recipes/DraconicItemsRecipe.java
@@ -27,7 +27,7 @@ import gregtech.common.metatileentities.MetaTileEntities;
 import gregicality.multiblocks.api.fluids.GCYMFluidStorageKeys;
 
 import gtexpert.api.GTEValues;
-import gtexpert.api.recipes.draconic.GTEDraconicRecipeMaps;
+import gtexpert.integration.deda.recipemaps.GTEDraconicRecipeMaps;
 import gtexpert.api.unification.material.GTEMaterials;
 import gtexpert.api.util.GTEUtility;
 import gtexpert.integration.gt.GTHelper;
