+++ src/main/java/gtexpert/integration/eio/EnderIOModule.java
@@ -25,7 +25,7 @@ import gregtech.api.recipes.ingredients.nbtmatch.NBTTagType;
 
 import gtexpert.api.GTEValues;
 import gtexpert.api.modules.GTEModule;
-import gtexpert.api.recipes.GTERecipeMaps;
+import gtexpert.core.GTERecipeMaps;
 import gtexpert.api.util.GTEUtility;
 import gtexpert.integration.GTEIntegrationSubmodule;
 import gtexpert.integration.eio.recipes.*;
